misc/HtmlTestSuite/results/test-0688.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head><title>CSS Flexbox: align-self: center content with flex-direction: column.</title>
<link href="https://drafts.csswg.org/css-flexbox/#flex-direction-property" rel="help"/>
<link href="https://drafts.csswg.org/css-align/#align-self-property" rel="help"/>
<link href="https://crbug.com/750553" rel="help"/>
<meta content="This test ensures that no unnecessary horizontal offset is applied to inline content in a 'align-self: center' box, inside of a flexbox with 'flex-direction: column'." name="assert"/>

<style>
html, body {
  margin: 0;
}

body {
  display: flex;
  flex-direction: column;
}

.content {
  align-self: center;
}

.content &gt; div {
  width: 400px;
  display: inline-block;
}
</style>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

</head><body onload="checkLayout('.content')">

<div class="content" data-offset-x="0">
  <div data-offset-x="0">X</div>
  <div>X</div>
  <div>X</div>
  <div>X</div>
  <div>X</div>
</div>

</body>
</html>
